requirements.txt
asgiref==3.8.1
Django==5.1.5
django-allauth==65.3.1
djangorestframework==3.15.2
et_xmlfile==2.0.0
openpyxl==3.1.5
pip==25.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.1
